Merge business-dev to 2.6-test

+++ seatunnel-connectors-v2/connector-s3-redshift/src/main/java/org/apache/seatunnel/connectors/seatunnel/redshift/commit/S3RedshiftSinkAggregatedCommitter.java
@@ -351,7 +351,9 @@ public class S3RedshiftSinkAggregatedCommitter extends FileSinkAggregatedCommitt
             String filepath = tempFilePath;
             mergeS3FileToRedshiftWithTemporaryTable(filepath);
 
-            fileSystemUtils.deleteFile(filepath);
+            if (filepath != null) {
+                fileSystemUtils.deleteFile(filepath);
+            }
             log.info("delete file {} ", filepath);
         }
     }
